LLM_Lab_Prompt_Engineering/requirements.txt
google-generativeai>=0.4.0
pyyaml>=6.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
numpy>=1.24.0
python-dotenv>=1.0.0  # Optional: for .env file support
